Trait pallet_bonded_coins_runtime_api::BondedCurrency

pub trait BondedCurrency<Block: BlockT, Balance, PoolId, Operation, AccountId, BondedAssetId, CollateralAssetId, BitType, Error>: Core<Block>where
    Balance: Codec,
    PoolId: Codec,
    Operation: Codec,
    AccountId: Codec,
    BondedAssetId: Codec,
    CollateralAssetId: Codec,
    BitType: Codec,
    Error: Codec,{
    // Provided methods
    fn quote_for(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        pool_id: PoolId,
        currency_idx: u8,
        operation: Operation
    ) -> Result<Result<Balance, Error>, ApiError> { ... }
    fn quote_for_low_and_high_bounds(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        pool_id: PoolId,
        currency_idx: u8,
        low: Balance,
        high: Balance,
        rounding: Round
    ) -> Result<Result<Balance, Error>, ApiError> { ... }
    fn query_pools_by_manager(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        account: AccountId
    ) -> Result<Vec<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>>, ApiError> { ... }
    fn query_pools_by_owner(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        account: AccountId
    ) -> Result<Vec<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>>, ApiError> { ... }
    fn encode_curve_coefficient(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        coefficient: String
    ) -> Result<Result<Coefficient<BitType>, Error>, ApiError> { ... }
    fn decode_curve_coefficient(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        bit_representation: BitType
    ) -> Result<Result<String, Error>, ApiError> { ... }
    fn pool_info(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        pool_id: PoolId
    ) -> Result<Result<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>, Error>, ApiError> { ... }
    fn pool_infos(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        pool_ids: Vec<PoolId>
    ) -> Result<Result<Vec<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>>, Error>, ApiError> { ... }
}

Runtime API to compute the collateral for a given amount and pool ID and to query all pool IDs where the given account is the manager or owner.

Provided Methods

fn quote_for( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, pool_id: PoolId, currency_idx: u8, operation: Operation ) -> Result<Result<Balance, Error>, ApiError>

Calculates the collateral for the given amount. The operation is determining whether the amount is minted or burned. The calculated collateral amount is based on the current state of the pool.

fn quote_for_low_and_high_bounds( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, pool_id: PoolId, currency_idx: u8, low: Balance, high: Balance, rounding: Round ) -> Result<Result<Balance, Error>, ApiError>

Calculates the collateral for the given integral bounds lower and upper. This function computes the collateral amount based on the provided lower and upper bounds, regardless of the current state.

fn query_pools_by_manager( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, account: AccountId ) -> Result<Vec<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>>, ApiError>

Query all pool IDs where the given account is the manager.

fn query_pools_by_owner( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, account: AccountId ) -> Result<Vec<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>>, ApiError>

Query all pool IDs where the given account is the owner.

fn encode_curve_coefficient( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, coefficient: String ) -> Result<Result<Coefficient<BitType>, Error>, ApiError>

Calculates the bit representation for the coefficient. The coefficient is constructed by coefficient_int.coefficient_frac. The first value in the tuple is the internal calculated coefficient, represented as a string. The second value is the bit representation.

fn decode_curve_coefficient( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, bit_representation: BitType ) -> Result<Result<String, Error>, ApiError>

Parses the bit representation for the coefficient to a human readable format.

fn pool_info( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, pool_id: PoolId ) -> Result<Result<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>, Error>, ApiError>

Query the pool status in a human readable format.

fn pool_infos( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, pool_ids: Vec<PoolId> ) -> Result<Result<Vec<PoolDetailsOf<AccountId, Balance, BondedAssetId, CollateralAssetId>>, Error>, ApiError>

Query the pools status in a human readable format.