Type Alias pallet_bonded_coins_runtime_api::PoolDetailsOf

pub type PoolDetailsOf<AccountId, Balance, AssetId, CollateralAssetId> = PoolDetails<AccountId, AccountId, CurveOf, CurrenciesOf<AssetId, Balance>, CollateralDetails<CollateralAssetId>, Balance, Balance>;

Human readable pool details.

Aliased Type

struct PoolDetailsOf<AccountId, Balance, AssetId, CollateralAssetId> {
    pub id: AccountId,
    pub owner: AccountId,
    pub manager: Option<AccountId>,
    pub curve: Curve<String>,
    pub collateral: CollateralDetails<CollateralAssetId>,
    pub bonded_currencies: Vec<BondedCurrencyDetails<AssetId, Balance>, Global>,
    pub state: PoolStatus<Locks>,
    pub currencies_settings: BondedCurrenciesSettings<Balance>,
    pub deposit: Balance,
}

Fields
id: AccountId

The ID of the pool.

owner: AccountId

The owner of the pool.

manager: Option<AccountId>

The manager of the pool. If a manager is set, the pool is permissioned.

curve: Curve<String>

The curve of the pool.

collateral: CollateralDetails<CollateralAssetId>

The collateral currency of the pool.

bonded_currencies: Vec<BondedCurrencyDetails<AssetId, Balance>, Global>

The bonded currencies of the pool.

state: PoolStatus<Locks>

The status of the pool.

currencies_settings: BondedCurrenciesSettings<Balance>

Shared settings of the currencies in the pool.

deposit: Balance

The deposit to be returned upon destruction of this pool.